Fix the missing cells in ArcGIS Pro
It has come to our attention that when data is stored in the int8 format, certain cells with values close to "zero" are being omitted in ArcGIS Pro. In the data packing algorithm, the near-zero value is encoded as -127. It is suspected that ArcGIS Pro mistakenly decode the -127 value as -128, which represents NaN (Not a Number), causing it to not appear in ArcGIS Pro.

To address this issue, the proposed fix aims to preserve both -128 and -127 values, while allowing a range of -126 to 127 for data packing purposes.